compensate for IO overhead (subprocess and UDP stats reading)
Currently omxplayer prints its stats over 400 times a second
to the console. Thus the pexpect call in read_position_local will be
tirggered equally often and equal amount of packages is sent over
the wire. As a result, position_local will be slightly behind.
The slave has slightly larger overhead because it additionally receives
network traffic in a seperate thread.